backend/requirements.txt
fastapi>=0.100
uvicorn>=0.20
torch>=2.0
torchvision>=0.15
pillow
numpy
tqdm
python-multipart
opencv-python
matplotlib